[InfoHud] Add current input size to the playback index
Currently, the playback index only shows the current tick of the recording.
Now a second variable is added to show the size of the recording with the format 0/0,
with the second one being the size.

To get the size, I use PlaybackController#size()-1 as the index is always one smaller than the size.
And to not get negative numbers when no inputs are present, I use Math.max(size, 0)